add Renderer project and Renderer class for OpenGL calls
class Renderer is a singleton class that handles all OpenGL calls. It's
in a separate Renderer project that compiles to a static library which
is then used inside PecanEditor to call render commands.

Also created a new Core project for now containing only PecanLogger.
It will contain all logic which is independent from OpenGL and Qt and is
essential to the engine itself.