add Config class holding data configurable from a config file
Config class contains functionality for reading a config file, which is a normal text
file where each line assigns a value to some configurable variable in
Pecan, for example buttonFont is such a variable that we support in our
Config so inside the config file there could be a line that says
"buttonFont=("Calibri Light",20)". This line will be read by the Config
class and the value ("Calibri Light",20) will be assigned to a static
variable Config::buttonFont which can be used anywhere in Pecan.

The purpose of this config is to have some user-configurable data that
can be easily changed without rebuilding the whole Pecan

All places in code where a magic constant was previously used are now
changed to use the value from Config